[zendesk] Add a 'default group' option
With this change, it is now possible to ensure that the correct Zendesk
group will be used for the assignment. If the assignee is not member of
this "default/onduty" group (or if this parameter is not set), then the
user's own default group will be used instead.

Fix GH-25